+++ src/main/java/net/thesquire/backroomsmod/util/ModUtils.java
@@ -7,7 +7,10 @@ import net.minecraft.block.Blocks;
 import net.minecraft.block.entity.StructureBlockBlockEntity;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtElement;
-import net.minecraft.registry.*;
+import net.minecraft.registry.Registries;
+import net.minecraft.registry.Registry;
+import net.minecraft.registry.RegistryKey;
+import net.minecraft.registry.RegistryKeys;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.registry.entry.RegistryEntryList;
 import net.minecraft.server.world.ServerWorld;
@@ -19,11 +22,12 @@ import net.minecraft.structure.StructureTemplateManager;
 import net.minecraft.util.*;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
+import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.math.Vec3i;
 import net.minecraft.world.World;
 import net.minecraft.world.gen.structure.Structure;
 import net.thesquire.backroomsmod.BackroomsMod;
 import net.thesquire.backroomsmod.config.ModConfig;
-import net.thesquire.backroomsmod.world.structure.ModStructureKeys;
 
 import java.util.Optional;
 
@@ -44,8 +48,30 @@ public class ModUtils {
         return pos;
     }
 
-    public static int boolToInt(boolean b, int first, int second) {
-        return b ? first : second;
+    public static int boolToInt(boolean b, int tr, int fl) {
+        return b ? tr : fl;
+    }
+
+    public static float boolToFloat(boolean b, float tr, float fl) { return b ? tr : fl; }
+
+    public static Vec3d vec3itod(Vec3i vec) {
+        return new Vec3d(
+                vec.getX(),
+                vec.getY(),
+                vec.getZ()
+        );
+    }
+
+    public static Vec3i getVec3iComponents(NbtCompound nbt, String name, Vec3i oldVec) {
+        return new Vec3i(
+                nbt.contains(name + "X") ? nbt.getInt(name + "X") : oldVec.getX(),
+                nbt.contains(name + "Y") ? nbt.getInt(name + "Y") : oldVec.getY(),
+                nbt.contains(name + "Z") ? nbt.getInt(name + "Z") : oldVec.getZ()
+        );
+    }
+
+    public static RegistryKey<Structure> structureIdToKey(String str) {
+        return RegistryKey.of(RegistryKeys.STRUCTURE, new Identifier(str));
     }
 
     // this is a slightly modified version of Minecraft's NbtHelper::toBlockState method